# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vending machine management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Operators ─────────────────────────────────────────────────────────────
  /operators:
    get:
      operationId: listOperators
      tags: [operators]
      summary: List all operators
      responses:
        "200":
          description: List of operators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Operator"
    post:
      operationId: createOperator
      tags: [operators]
      summary: Create an operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
  /operators/{id}:
    get:
      operationId: getOperator
      tags: [operators]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Operator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
    patch:
      operationId: updateOperator
      tags: [operators]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
    delete:
      operationId: deleteOperator
      tags: [operators]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Stock ─────────────────────────────────────────────────────────────────
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List service schedules
      responses:
        "200":
          description: Schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a schedule entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /schedules/execution:
    get:
      operationId: getScheduleExecution
      tags: [schedules]
      summary: Get schedule execution status for current week
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Schedule execution status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleExecution"

  # ── Expenses ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [coffee, vending]
        purchasePrice:
          type: number
        salePrice:
          type: number
        unit:
          type: string
        stockQuantity:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - purchasePrice
        - salePrice
        - unit
        - stockQuantity
        - createdAt
    ProductInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [coffee, vending]
        purchasePrice:
          type: number
        salePrice:
          type: number
        unit:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - name
        - type
        - purchasePrice
        - salePrice
        - unit
    ProductUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        purchasePrice:
          type: ["number", "null"]
        salePrice:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Operator ─────────────────────────────────────────────────────────────
    Operator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [operator, moderator, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - isActive
        - createdAt
    OperatorInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [operator, moderator, admin]
        password:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - name
        - type
        - isActive
    OperatorUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]

    # ── Stock ─────────────────────────────────────────────────────────────────
...
    Schedule:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        dayOfWeek:
          type: integer
          description: "0=Sunday, 1=Monday, ... 6=Saturday"
        createdAt:
          type: string
      required:
        - id
        - clientId
        - clientName
        - dayOfWeek
        - createdAt
    ScheduleInput:
      type: object
      properties:
        clientId:
          type: integer
        dayOfWeek:
          type: integer
      required:
        - clientId
        - dayOfWeek
    ScheduleExecution:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        dayOfWeek:
          type: integer
        scheduledDate:
          type: string
        executed:
          type: boolean
        loadCount:
          type: integer
      required:
        - clientId
        - clientName
        - dayOfWeek
        - scheduledDate
        - executed
        - loadCount

    # ── Expenses ─────────────────────────────────────────────────────────────
...