# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vending machine management API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputType } from './productInputType';

export interface ProductInput {
  name: string;
  type: ProductInputType;
  purchasePrice: number;
  salePrice: number;
  unit: string;
  /** @nullable */
  notes?: string | null;
}